room_management_svc/requirements.txt
# Format
black == 24.8.0
flake8 == 7.1.1 
isort == 5.13.2 

# Log
loguru == 0.7.2 

# Test
pytest == 8.3.2 

# gRPC
grpcio == 1.65.4 
grpcio-tools == 1.65.4 

# Datamodel
pydantic == 2.8.2 
pydantic-settings == 2.4.0 

# Database
psycopg2-binary == 2.9.9 

# Migration
yoyo-migrations == 9.0.0 

# Dotenv
python-dotenv == 1.0.1 

# Redis
redis == 5.0.8 
